All shells export modified variables that have previously been exported.
Mac OS X 11 Big Sur ships with the widest variety of POSIX-compatible shells ever: bash, dash, and ksh. All three of them mark a variable for export when inherited from the parent environment or when used with the 'export' built-in. Once marked for export, only unsetting the variable clears the export bit.